updated the example of rewrites to utilize the App Router
