+++ examples/rewrites/app/_components/Code.tsx
@@ -1,4 +1,4 @@
-import styles from "../styles.module.css";
+import styles from "../../styles.module.css";
 
 type CodeProps = {
   children: React.ReactNode;
